HBASE-16970 Clarify Scan comment about caching
The Scan class has an integer field, caching, that defaults to -1.
The meaning of -1 is that the option was not overriden (aka not set) and
therefore that the default configuration should be used.

This trivial patch corrects the comment of that field to reflect this
meaning rather than misleading the reader to believe that caching is
disabled.